Added new workflows
build_and_publish_pypi.yml is needed to solve the catch 22 of not being
able to build the AppImage if it refers to the OSCR-UI package on PyPi
as a dependency, in the case that the wheel on PyPi is broken.

This is a manual upload to PyPi.

manual_build.yml is the same as build.yml but with a manual trigger
rather than a automatic on new release trigger.