Added an input parameter for the tag
Since the workflow condition is replaced by a manual dispatch, there
will be no tag for this condition to trigger.

```
if: startsWith(github.ref, 'refs/tags/v')
```

The version tag needs to be a workflow parameter, and the condition
needs to refer to that input parameter.

And there won't be a tag under 'refs/tags/v' because it won't be created
yet, we're creating a release with the tag from the workflow input.

Therefore the condition has to be changed, checking if inputs.tag starts
with 'v', a version number.