Added two experimental workflows for building with PyInstaller
These only produce artifacts that can then be downloaded from GitHub for
the moment. I am unsure whether or not they will work, as there are Qt
dependencies that might be unaccounted for in the Git VM's .venv

After some more feedback and testing, perhaps missing libraries that are
unaccounted for will make themselves known through trial and error.