Additional fixes to the Inno .ISS file
I forgot that dist contains a subfolder of the project name and not the
direct .exe and _internal folder.

Also the output base name would conflict with the actual binary, so
added Installer as a suffix.

And WizardStyle refers to the dialog when installing/uninstalling.
Not the script generation wizard.