Fixed both workflow files
${{ matrix.runner }} was missing a $
(it's to be able to define several operating systems for one workflow to
run undere, for eventual production I'll use the workflow I'll write and
submit later to unify Linux and Python binary compilation)

Additionall, the shell: was in the wrong place.

Removed lien separation for packages entirely because it's just another
source where it makes the script less transferrable to a Windows
workflow (which is what I will do later, Windows+Linux build in one
workflow.