Added artifact extensions for impending merge
Ensuring that nothing makes its way onto the main branch, no build
artifacts, nothing that should't just e sitting there.

However.. I might remove *.desktop from that list, depending on how I
structure the build process. In the end it's just a tiny INI file that
can be generated upon installation. Does make it more flexible taking
versions into account, since if the *.desktop entry is generated by the
PKGBUILD (which is just a bash script at the end of the day) then it
means it'll always contain the accurate metadata, and we do not have to
maintain it independently.