First batch of workflow structure improvements
This is a reimplementation of the PyInstaller workflows that build the
Linux and Windows binaries respectively; if these both work on GitHub,
they could be merged into one theoretically. See the difft screensohot
(will attacho once on GitHub) for a better diff of precisely what was
standardized between both. (look for not just blue/red, but underlines
characters and words)