Replaced wildcard glob copy with explicit copy.
The "*" symbol relies on a shell to perform expansion, and while the
PKGBUILD file is indeed a bash script, it is not run through /bin/bash
rather makepkg interprets the bash source code and executes it as just
that: code. Or put another way, it treats the content of the file as
the final state, except for string interpolation and variables that it
defines before running it.

It does not perform shell expansion, so the * character is taken as a
literal file name. This is the behavior I observed. There may be a way
to do globbing in a PKGBUILD some other way but this is not it. There
are only two things to copy anyway, so here I replaced it with just
explicit copies.